added agents, movement, contact
A green cube (target) was added to the environment.

A white sphere representing a sphero was added with functionality allowing it to move around the environment with user input through ASDW keys, as well as handling contact between itself and the green cube.

A spotlight agent was added with functionality allowing it to move around the environment with user input through IJKL keys.