mapped contact and added shapes to environment
This new version has many more shapes in the environment, both the green target cubes and other non-target shapes.

When the sphero contacts anything in the environment, the location is stored, and the contacts are mapped at the end of the simulation.

The sphero is now controlled by the keys 2qwe, because the previous controls included the s key, which Mujoco uses as the key to make the simulation run slower.